added confirmation popup before deleting a todo and added a message which will be displayed when there is no data found after applying a filter

+++ src/app/todos/todo-list/todo-item/todo-item.component.ts
@@ -25,6 +25,8 @@ export class TodoItemComponent implements OnInit {
   }
 
   onDeleteTodo(todoId:string){
-    this.todoService.deleteTodo(todoId);
+    if(confirm("Are you sure you want to delete this item?")){
+      this.todoService.deleteTodo(todoId);
+    }
   }
 }
